simplify topicPartitionsData
There were multiple TODOs around just keeping a number of partitions; we
can simplify that further by just keeping the partition data directly in
a contiguous slice.

With the previous validation added to metadata that enforces the
response contains partitions in order from 0 with no gaps, we can use
that response blindly.

This removes two maps and simplifies the topic partitions structure, as
well as removes []int32 all over in favor of just a direct count.