kversion: mild redux (and breaking) redux; introduce VersionGuess
Breaking: everything is converted to returning *Versions rather than
Versions. This allows for future growth of the Versions struct.

Redux: version slices are now defined once per version, and these
versions are duplicated in responses. This is different from before,
where we built a version up from scratch using the prior version,
recursively from 0.8.0. This is cleaner, albeit requires more
duplication of the same variables, which unfortunately cannot just be
abstracted away into the future (one definition, one use in the method,
one use in VersionGuess, one use in defining the next version).

VersionGuess is a handy function to analyze the versions and guess which
Kafka version we are talking to. This does more work than expected to be
as precise as possible.