kgo: allow -2 and -1 in Offset.At
@weeco ran into a problem with this. The original reason for bounding to
-2 (the start) when seeing a negative was to address the lazy behavior
where somebody just passed negatives to consume at the beginning.

This change basically allows one new special offset, -1 (since -2 and
any other negative was already bounded to the start). There is a small
risk that a person not reading the function signature docs will try -1
and unexpectedly not be at the start, but ideally, the method will be
followed as specified.

This is technically non-breaking since before, we did not document the
behavior around negative offsets. The prior unwritten assumption was
that only valid offsets were used.