kversion: remove patch from VersionGuess when we cannot know it
Pre 1.0.0, patch numbers actually went v0.9.0.1; we already did not
include the final .1. Post 1.0.0, we cannot know patch numbers, so let's
exclude them.

Also changes from saying "tip" to "at least <final verison kversion
knows about>", so a user can know how out of date their kversion / kgo
is.

Also moves the cmp map to the top to only allocate it once. We either
clear it every loop, or it is non-empty which results in the "under"
case which returns within that loop.