producer: reset all partition's sequence numbers for idempotent producer OOOSN
See embedded doc or mentioned KAFKA ticket.

This adds resetting the batch index to resetSeq. Historically, this was
not strictly necessary, but it's a good idea anyway. Historically, a
resetSeq on OOOSN also added the batch to reqRetry, which eventually
reset the batch drain index. The other resetSeq was in txn once
everything was aborted and nothing was producing, so the batch drain
index was 0. The new logic for the KAFKA ticket resets all partition
sequence numbers during production while nothing is stopped; thus we
need to absolutely ensure the batch drain index is 0 while resetting.