kgo: fix three races
One of these is new and unreleased,
one was introduced in v0.6.2,
and one has apparently existed since last summer and never cropped up?!?

The unreleased one is fairly obvious: we modify bufferedFetch when
closing a consumer session, so our defer cannot touch it.

The v0.6.2 one was rather obvious as well. The AND logic caused us to
fall into a branch we should not have. We instead move the AND into a
block on its own.

The one since last summer does not make sense and was never seen in
races before. I also don't think this is a franz-go problem but may
instead be a compress one hanging onto the input slice. Basically, the
race detector was showing that the basic logger was reusing a slice that
was currently in use in the zstd decompressor, which should not be a
thing since we get and put back into the slice writers pool in defined
blocks. The decompressor is **done** with the input slice on return!

But, as it turns out, resetting the reader was literally useless, we do
not need to do that at all, so we just don't.