consuming: bugfix for consuming compacted topics
See KAFKA-5443: a compacted topic can have later records removed from
within a batch, but the last offset delta is preserved. We need to use
the last offset delta for a batch rather than just relying on consumed
records. If we do not, we could ask for 144, get 143 while having a last
offset delta of 145, and never advance since we discard 143 and do
nothing about 145.