kgo: fix NoAck produce requests
Kafka sends **no response** when we produce with no acks, so the current
code was wrong: we would send a request and then it would effectively
hang until the read timeout. Then we would re-send the request.

This now adds the proper shortcut logic such that as soon as we write
the request successfull, we finish all promises.

We still retry on error because if we receive an error, we did a partial
write and kafka did not understand our request.