consumer group: fix race
The the group was rotated, we did not wait for the revoke to be done
before falling back into managing.

Revoking happens in a dedicated goroutine, and managing revokes as well
on return (different type of revoke). These two revokes raced.

We now unconditionally wait for the heartbeat function to finish
revoking.