group balancer: debug => info logging; handle join better
Previously, if a bunch of group members joined with disparate interests,
the leader would only know how to assign topics that it was interested
in, not what everything else was interested in. Now, if members join
with topics the leader does not know of, the leader performs a metadata
fetch on all topics.

However, we still do not keep around information of these other
interests, meaning that if an admin adds partitions to a topic later,
the leader will not notice this. The risk here is pretty low, but if
this is an issue later, we can add an option to keep the metadata for
all topics. The downside of this is that all leaders eventually get
bloated topics lists, meaning metadata requests slowly grow larger over
time.

With this, we also change up the logging messages a bit to be
potentially nicer, and we move them to the info level so that we always
log. For large balances, this could be a really, really big message, but
ideally balances are pretty rare.