consumer: properly shut down manageFetchConcurrency
There were two issues.

First, we only checked context cancelation when we received a response.
For the most part this was fine, but if a session shut down while a
source was not wanting a fetch, the goroutine would leak. The fix is to
always select on context cancelation and if we see it is canceled, quit
as soon as we hit zero active fetches. This allows for us to see if a
session is quit even if nothing is actively fetching.

The second problem was that a source could indicate it wants to fetch
and then return after the indication. The manager would reply back in
the 1buffer chan that the source could fetch and then bump its active
fetches, but the source had already returned. Thus, active fetches would
never hit zero. The fix is in the source to cancel its desire.