sticky: fix and drastically simplify isComplex detection
Before this commit, if a sticky balance with tons of disparate members
only passed in **one** topic, then the code was unable to detect if the
balance was complex, because we only detected complexity after the first
topic.

Now, we simply check if all members want to consume each topic equally.
topics2memberNums is built from topics each member is interested in, so
if any topic one member is interested in is not interested in by all
members, then we know we are complex.