consumer group: minor code cleanup
This moves non-cooperative assignment invalidating into the revoke code
directly, replacing multiple comments that refer to when things happen.
This simplifies the reasoning for assignInvalidateAll when returning
from error in the manage goroutine as well.

As well, this uses the group context rather than a background context
for autocommitting.

Finally, this moves the channel creation / defer wait to _after_ the log
defer wait so that the log is printed after the wait, not before.