kversion: support guessing with skipped keys, guessing raft
A few months ago, Kafka added a "listeners" section to the json
definitions of requests. This section corresponded to which listeners
could handle those requests: the classic zookeeper broker listeners, the
new raft brokers, or the raft controller broker.

This commit ports the listeners over to kversions and adds support for
guessing the version when speaking to a raft broker.

The initial attempt was to try guessing all kinds of brokers at once,
but this quickly gets pretty weird, and I think it would be easier as a
user of VersionGuess to try againt zk / raft as desired.

This also adds support for skipping keys and defaults to skipping the
four broker-to-broker keys. Oracle does not provide these four
broker-only keys, and neither does redpanda. They aren't really
meaningful in a version guess, so we may as well default to skipping
them, but allow for not skipping them by providing an empty SkipTags
option.