group consumer: revoke all on LeaveGroup, properly blocking commit
Leaving the group cancels the group context. This context is used in
defaultRevoke, meaning that leaving a group prevents an offset commit in
onRevoke to be successful. Now, we use the client context, which is only
canceled after unsetting a consumer, which waits for the group to be
left.

Additionally, we now revoke all offsets if cooperative if we are leaving
the group. Previously, we would rely on the manage logic to call
onRevoke, but this also sometimes calls onLost. Technically with onLost,
it is "too late" to commit at that point. We do not want to imply that
if a user specifies their own OnLost.

Finally, we default onLost to defaultRevoke. We implicitly did this
before by falling into onRevoke if onLost was nil, but we may as well
explicitly do this.