source: more properly ignore truncated partitions
A truncated partition can come at any moment, even before we read the
magic. We need to not set the partition error if we fail reading. In
some places, particularly when we validate we have enough length ahead
of time, we do now properly set the partition error because if we have
enough length, we should be able to decode.

We now only use LastOffsetDelta if we properly decoded all records. If
we did not before, then we could have failed on reading a truncated
batch and then skipped all records.

Finally, if we decode a magic we do not know of, we set an error in our
response and properly skip the offset.