sink: allow concurrent produces
When doSequenced from the broker to the sink, I accidentally lost the
concurrent issuing aspect. The sequenced aspect in the sink issued and
processed requests in order, but using waitResp meant we actually never
issued more than one request at a time.

Thix fixes that by ensuring we issue the produce with `do` before we go
to waiting for a response.