sink&co: beef up debug logs
This drastically beefs up debug logs for producing so that we can see
what partitions are being produced to, at what offsets, what the final
offset is, and if any retriable or fatal errors occur.

This also reliably adds "broker" to all source / sink / broker logs.

Lastly, this "simplifies" the handleReqResp loop by breaking out the
core batch processing logic into a dedicated function. Rather than
checking isFirstBatch and then checking again when finishing the batch,
we now just grab the recBuf lock always with the assumption that we are
usually finishing the batch successfully, so our new lock usage is
actually simpler and more efficient.