breaking API: error redux
This makes private most errors; it was not really useful to have them
public. Only a few remain public: ErrDataLoss, ErrMaxBuffered, and
ErrAborting. These could likely be internal, but may be useful to check
specifically sometimes. I could imagine an alert being setup on
ErrDataLoss.

Some errors are now gone and replaced with more detail with an
fmt.Errorf at their original callsites. Some errors that were previously
retriable are now deemed not retriable; in these instances, we were
retrying to work around bad Kafka behavior (small responses, etc).
Really, we should be talking to working Kafka's.

ErrConnDead is overhauled to only be returned on read or write errors
now. It was previously used all over, but a lot of the instances where
it was used were either papering over a broken Kafka, or the wrong
error (e.g., should have been errChosenBrokerDead).

This also now no longer wraps the dial error. We return it blindly and
trust that it will implement Temporary() properly.

Overall, this commit was a long time coming and should help with some
mysterious retrying.