broker: read the cxn even when not having a request
Previously, we would only read the connection if we had written a
request. What would often happen to the "normal" connection is that
Kafka would close it due to it being idle, and then we would only notice
when we went to write again.

By always reading, we can notice this connection closure as Kafka closes
us and can avoid an unnecessary errDeadConn on the first write.

This does unfortunately make the code a good bit more complicated.