broker: commit what we are imminently aborting
For posterity, this is more progress on the attempt to always read
connections. I've realized this will not interact well with our sasl
flow, as well, this has a bug that is quite difficult to trace down.
Lastly, we do not actually need this for connection reaping, so we may
as well just rely on our own idle timeout and connection reaping and
avoid the complication and ugliness that the read conn async stuff
introduced.