broker: revert always reading the connection
As mentioned in the immediately prior commit, this had some
unexplainable bugs, and after realising that the sasl token refresh
would not play nicely AND that idle conn reaping suffices most of the
same purpose, I'm less interested in trying to track down what's up,
especially since no matter what's done, the read flow itself was uglier
and slower (through two channels).

So, in short, let's just keep things simple and rely on idle conn
reaping.

The main problem we were trying to prevent by always reading the conn is
to detect kafka closing the connection when we haven't written for a
long time. Not detecting it meant that connections sat in a dead zombie
state in the kernel, and when we did eventually write, we would
immediately error.

With connection reaping, we should detect these idle connections before
Kafka kills us, so we should not have dead sockets in the kernel and we
should not write&error immediately. And, even if we still experience the
write&error occasionally, we can always reinit the connection easily.