generate: support parsing and encoding unknown tags
For parsing, we simply stop using SkipTags and now use ReadTags to set
the new UnknownTags field that is in all flexible structs.

For appending, we need to switch to using reflect.DeepEqual when the tag
is a struct, since a map is now in the struct. This *should* be fine
since the struct is simple, minus the tags. As well, we no longer just
append 0 tags blindly; instead, we always have to append unknown tags as
well.