producer: retry logic changes
producerID() could return errReloadProducerID, which would fail all
buffered records. Instead, we want to bump all records load errors by 1,
because this is actually a compeltely retriable error.

AddPartitionsToTxn was permanently looped, but we can use this same
bump-and-retry logic to be more consistent and remove an infinite loop.

---

When addressing KAFKA-12671, I put the lastRespSuccessful in the wrong
spot. It's not moved to a more obviously right spot. The addressing was
still successful in integration testing because we still waited for all
produce responses to finish before issuing EndTxn, and in integration
tests, all requests are successful (no request failures).

I also accidentally used incDrains/decDrains instead of
incIssues/decIssues. This was non-impactful, as both drains and issues
are waited on when aborting, but it was less clear. It may be worth it
to just merge them anyway, and I may do that soon.

---

Finally, we should set the initial failing state for a recBuf to whether
there is an ErrorCode. This will help us avoid trying to produce if we
know we loaded an error, which may make what follows this easier.