consumer: switchup how topics are stored
This addresses one big part of #30: if we want to be able to delete
topics that fail loading for a long time, it is much easier to reason
about the safety of this if what we are consuming is stored separately.

We now store what topics we are consuming directly in the groupConsumer
or directConsumer structs.

This actually makes some things easier to reason about--rather than
loading topics and then reloading them later only to have maybe changed,
we wire through what we are consuming into the consumerSession and use
that wherever relevant. If the consuming changes, the session changes,
and we get to worry less about some weird map mismatch stuff.

The client's topics map now corresponds to just producer topics. The
next commit will move it to the producer struct directly.

A few other small changes:

- doOnMetadataUpdate is now unconditionally called at the end of
  updateMetadata. Previously, we would not call it if the request
errored, which had the potential of leaving a consumer session's
listOrEpochLoad hanging. Now, if the request fails, the list or epoch
load will proceed only to itself fail later and reload.

- we no longer immediately trigger a metadata update at the end of
  assignPartitions, because all assignPartitions are triggered after a
metadata update.

- we bump the min metadata trigger window from 1s to 2.5s just to ensure
  we stop being so spammy