consumer: remove cursors from source when unsetting
We now no longer have a strictly append operation on an existing map.
We can now completely replace old cursors.

This means that when we reassign and have a completely new
topicsPartitions, we abandon old cursors. Before this, commit, they
would hang around.

It would be unsafe simply to remove all cursors from sinks outside of
the context of the metadata loop, because a metadata could have loaded a
topicsPartitions and can undo any removals. Instead, we need to
introduce a mechanism to hook into the metadata loop outside the context
of an update.

This does mean now that a reassignment may wait a little bit for an
active metadata request to complete. Realistically, that should not wait
long.