producer: allow using the passed in context to cancel records
Building off the prior commits, before we add a partition to a batch, we
will check if the first record is canceled. If so, we fail the entire
partition.

We fail an entire partition rather than just the first batch because it
is (a) easier, and (b) we really should not be taking so long, so if
producing the first record takes a while, there's a problem.