producing: evaluate whether a batch should fail before and after
We were inconsistent on evaluating whether a batch should fail. Before
writing a request, we would check if a record ctx expired. After, we
would evaluate the record timeout and the produce retries. We may as
well evaluate all three at once both before and after for consistency.

We also now allow checking whether records should be failed before
writing a request in the non-idempotent context as well.