metadata: do not nil cursors/records pointers ever
If a user is consuming and producing to the same topic in the same client, then nilling either of these pointers will cause a panic.

This was an oversight I had while splitting the producer and consumer topic stores; I figured I could nil these pointers to save some memory that is hanging around but that actually is not the case. Overall, the memory savings would be pretty small and not meaningfully impactful, so not nilling avoids any problems.