sink: bugfix firstTimestamp
In the producer guts overhaul, I changed where firstTimestamp is
calculated. Unfortunately, calculateRecordNumbers now happens before
setting firstTimestamp, so the **first** record added to a batch
calculated a timestamp delta off of 0, which was large and would put the
batch creation time into the future. All other records would then
calculate a timestamp delta off of the actual real firstTimestamp,
because the first record would be buffered, and the buffering is what
sets firstTimestamp.

Now, in calculateRecordNumbers, if the record is to be the first record,
then we zero out tsDelta.