sink: create producerID *BEFORE* produce request (partial revert of dc44d10b)
dc44d10b was done because a metadata update would trigger a sink to
drain, and that would ALWAYS create a producer ID, and would then see
there was nothing to produce. So, we were creating producer IDs even for
pure consumers.

Unfortunately, switching the order meant that the errReloadProducerID /
needSeqReset combo was flawed. In fact, I even noted that the order
needed to be producerID and then createReq in the bottom of the docs
just above the producerID call.

We now re-move the producerID call above createReq, but we preserve the
intent of dc44d10b by checking if any records are buffered before
loading the producer ID.

This also adds an important doc comment with a big NOTE just above
createReq to prevent this order flipping from occurring again, as well
as makes the prior doc comment just above producerID more prominent.