producer: drop error return from Produce (breaking API change)
The early error return was useful in some ways, but was also a footgun:
there were only three possible errors that could be returned, and the
api was prone to a user deadlocking. For example, if a person added to a
WaitGroup only if Produce returned nil, then with pathological
scheduling, it would be possible for the WaitGroup to be decremented
through the promise before the nil error check occured. Such usage would
cause a panic.

The early error return also was tedious to deal with, and it was not so
obvious why sometimes Produce would return early vs. the majority of the
time when a promise would be called.

To alleviate these issues, we will just **always** go through the
promise. The main problem here is that we want to support the following
behavior:

    ch := make(chan error)
    cl.Produce(ctx, r, func(_ *kgo.Record, err error) {
            ch <- err
    })
    return <-ch

If we do not invoke early-return errors in goroutines, then this could
deadlock, because the promise would be sending to ch before a receive is
ready. So, to support this behavior, we invoke early return errors in
goroutines. Ideally this is a non-issue, because early return errors
usually signal the client is doing something not quite right.