txnal producer: work around KAFKA-12671, take 2
Rather than ever aborting anything when we could potentially be unsafe,
we will work with the logic we added to ensure we do not have broken
sequence numbers: only if we can abort, do we.

Secondly, we now wait for everything to be "flushed" (with the caveat
that now we can abort when flushing).

These two in combination should avoid KAFKA-12671.