txn: small retry / producer id fail / doc updates
When ending a transaction, we now document that only
OperationNotAttempted is retriable. All other errors really are
technically not, not even for retrying with an abort.

We now retry ending a transaction few extra times on
ConcurrentTransactions, because of reasons (see doc on
doWithConcurrenTransactions).

If our EndTxn or AddOffsetsToTxn fails, we now fail the producer ID if
we should, rather than allowing it to continue to be valid.

Lastly, we avoid failing all buffered records if AddPartitionsToTxn
fails. We will let users do this with EndTransaction. This helps avoid
potential weird KAFKA-12671 issues. The only remaining uses of
failBufferedRecords are when the client itself is closing.