consumer: fix SetOffsets bug
For assignSetMatching, we want to keep anything that is currently
loading, because anything currently loading either matches and needs to
continue loading, or does not match... and is not to be messed with,
i.e., should also continue loading.

I'm not sure why 2.8.0 finally cropped this up.

This also clarifies some documentation, and removes a fatal test quit
when the group cannot be deleted, because 2.8.0 also has some ... timing
issues ... so some things may take some time to delete. So long as all
the numbers in the integration tests line up, then there is no reason to
fail the test.