aws_msk_iam sasl: add support for session tokens, user agents
A session token is very easy to support, and is useful.

A user agent is not as easy to support in a great way. A go package has
no way to determine which tag of the package was used, so we only give
the module used, the go version, and the hostname.

I've tested this locally against a scrappy cluster, so ideally this will
work for others too.

This also drops the presign timeout down to 5min, because there is no
reason to have it be 15min. Even 5min is generous and I would rather do
1min, but maybe some extremely awful network needs 5 min to login.