hooks: add HookBrokerE2E
Previously, it was very difficult to track e2e information. This new
hook makes it possible and easy.

The only downside is that now we are passing around a lot more
arguments, but ultimately this is non-impactful in comparison to a
network write.

This hook can completely replace the write and write hooks, but it's
easy enough to keep those around, so might as well.