client: fix request buffer pool, readd pnr pool
Previously, we were not keeping the inflated buffer, we would put back
into our pool the original buffer. This fixes that, which drastically
reduces garbage for produce requests.

Secondly, after the multiple producer improvements, I'm more confident
in adding a pnr pool similar to the one that originally existed. The
only difference now is that the pnr pool is scoped per client, rather
than globally, so it is safer in the face of races. Again though, after
the previous reduxes, I'm more confident in the pool to begin with.